requirements.txt
# annotated-types==0.7.0
# anyio==4.5.2
# certifi==2025.8.3
# cffi==1.17.1
# charset-normalizer==3.4.3
# click==8.1.8
# cryptography==46.0.1
# distro==1.9.0
# fastapi==0.118.0
# h11==0.16.0
# httpcore==1.0.9
# httpx==0.28.1
# idna==3.10
# jiter==0.9.1
# openai==1.109.1
# pycparser==2.23
# pydantic==2.10.6
# pydantic_core==2.27.2
# PyGithub==2.8.1
# PyJWT==2.9.0
# PyNaCl==1.6.0
# python-dotenv==1.0.1
# requests==2.32.4
# sniffio==1.3.1
# starlette==0.44.0
# tqdm==4.67.1
# typing_extensions==4.13.2
# urllib3==2.2.3
# uvicorn==0.17.0
# google-generativeai==0.4.0


fastapi==0.118.0
uvicorn==0.17.0
python-dotenv==1.0.1
PyGithub==2.8.1
httpx==0.28.1
openai==1.109.1
google-generativeai>=0.5.0
requests==2.32.4
pydantic==2.10.6
